Add save session capability for tmux terminals
Tmux can be called automatically on login. The terminal should run
scripts/tmux-setup.sh to start a session properly. The script creates a
new session with the proper name (terminal or terminal-n). The script
also handles cleanup. On logout, if the tmux session is no longer
attached and hasn't been marked non-killable in ~/.tmux-saves/ the
script will kill the session.

To mark a tmux terminal session as non-killable, run bin/tmux-save
(aliased to tsave in aliases.sh). This creates a file with the name of
the tmux session in ~/.tmux-saves/.

The above only happen for tmux sessions that are started automatically
using tmux-setup.sh. Manually starting a tmux session will not have any
of this capability.

Aliases added:
- tkill: tmux kill-session
- tsave: tmux-save (~/bin/tmux-save)

+++ scripts/tmux-setup.sh
@@ -12,17 +12,20 @@ else
     SESSION_NAME="$BASE_SESSION"
 fi
 
+# Remove save file if exists
+rm -f ~/.tmux-saves/"$SESSION_NAME"
+
 # Start session (with message if there are detached sessions)
 if /usr/local/bin/tmux list-sessions 2> /dev/null | grep -vq '^.* (attached)$'; then
-    /usr/local/bin/tmux new -s "$SESSION_NAME" \; display-message "WARNING: detached sessions" > /dev/null 2>&1
+    /usr/local/bin/tmux new -s "$SESSION_NAME" \; setenv SESSION_NAME "$SESSION_NAME" \; display-message "WARNING: detached sessions" > /dev/null 2>&1
 else
-    /usr/local/bin/tmux new -s "$SESSION_NAME" > /dev/null 2>&1
+    /usr/local/bin/tmux new -s "$SESSION_NAME" \; setenv SESSION_NAME "$SESSION_NAME" > /dev/null 2>&1
 fi
 
-# After exiting, kill session if it is not attached
-if ! /usr/local/bin/tmux list-sessions | grep -q "^$SESSION_NAME: .* (attached)\$"; then
+# After exiting, kill session if it is not attached or liste in ~/.tmux-saves/
+if [ ! -f ~/.tmux-saves/"$SESSION_NAME" ] && ! /usr/local/bin/tmux list-sessions | grep -q "^$SESSION_NAME: .* (attached)\$"; then
     /usr/local/bin/tmux kill-session -t "$SESSION_NAME" > /dev/null 2>&1
-    echo "[kill-session $SESSION_NAME]"
+    echo "[kill session $SESSION_NAME]"
 else
     echo "[detach from $SESSION_NAME]"
 fi
